pyproject.toml
[tool.poetry]
name = "brewblox_brewfather_uploader"
version = "0.1.0"
description = "Brewblox service for uploading temperature and gravity information to Brewfather"
authors = ["Andrew Stephon <andrew@stephon.dev>"]
license = "GPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.7,<4"
brewblox-service = "^0.33.0"
pyyaml = ">=6.0"

[tool.poetry.dev-dependencies]
pytest-flake8 = "^1.0.7"
pytest-cov = "^2.11.1"
pytest-mock = "^3.5.1"
pytest-aiohttp = "^0.3.0"
flake8 = "^3.8.4"
autopep8 = "^1.5.5"
pytest = "^6.2.2"
flake8-quotes = "^3.2.0"
aresponses = "^2.1.4"
mock = "^4.0.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
